Improve round map (fixes #164)
