+++ app/javascript/redux/events/roundMap/designatedLane.js
@@ -1,5 +1,7 @@
 const prefix = '[events/roundMap/designatedLane]'
 
+import { LOAD_REQUEST } from './actionTypes'
+
 const UPDATE_DL = `${prefix} UPDATE`
 const RESET_DL = `${prefix} RESET`
 
@@ -28,14 +30,15 @@ export function updateDL({ enabled, startPoint, endPoint, laneViolation }) {
 
 export default function reducer(state = initialState, action = {}) {
   switch (action.type) {
+    case LOAD_REQUEST:
+    case RESET_DL:
+      return {
+        ...initialState
+      }
     case UPDATE_DL:
       return {
         ...state,
         ...action.payload
-      }
-    case RESET_DL:
-      return {
-        ...initialState
       }
     default:
       return state
